requirements.txt
Django==3.2.10
django-cities==0.6
django-cors-headers==3.10.1
django-extensions==3.1.3
djangorestframework==3.12.4
drf-exceptions-hog==0.1.0
drf-jwt==1.19.1
psycopg2-binary==2.9.2
social-auth-app-django==5.0.0